feat: improve card distribution #29
