+++ mindbug-backend/.gitattributes
@@ -1,3 +1,2 @@
 /mvnw text eol=lf
 *.cmd text eol=crlf
-* text=auto
