+++ mindbug-backend/.gitignore
@@ -30,4 +30,3 @@ build/
 !**/src/test/**/build/
 
 ### VS Code ###
-.vscode/
