+++ mindbug-backend/src/main/java/com/mindbug/controller/GameController.java
@@ -36,16 +36,16 @@ public class GameController {
 
     @PostMapping("/attack")
     public ResponseEntity<String>  attack(@RequestBody PlayerCardDto data) {
-        this.gameServer.handleAttack(data.getPlayerId(), data.getGameId(), data.getSessioncardId());
-        return ResponseEntity.ok().build();
+       this.gameServer.handleAttack(data.getPlayerId(), data.getGameId(), data.getSessioncardId());
+       return ResponseEntity.ok().build();
     }
 
     @PostMapping("/dont_block")
     public ResponseEntity<String>  dontBlock(@RequestBody PlayerIdDto data) {
-        this.gameServer.handleDontBlock(data.getPlayerId(), data.getGameId());
-        return ResponseEntity.ok().build();
+       this.gameServer.handleDontBlock(data.getPlayerId(), data.getGameId());
+       return ResponseEntity.ok().build();
     }
-
+    
     @PostMapping("/game/play_card")
     public ResponseEntity<String> playCard(@RequestBody PlayerCardDto data) {
         this.gameServer.handlePlayCard(data.getPlayerId(), data.getSessioncardId(), data.getGameId());
