+++ mindbug-backend/src/main/java/com/mindbug/dtos/PlayerBasicInfoDto.java
@@ -21,4 +21,7 @@ public class PlayerBasicInfoDto {
         this.playerId = player.getId();
         this.nickname = player.getNickname();
     }
+
+
+
 }
