+++ mindbug-backend/src/main/java/com/mindbug/models/Effect.java
@@ -1,89 +1,52 @@
 package com.mindbug.models;
 
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.ElementCollection;
+import jakarta.persistence.CollectionTable;
+import jakarta.persistence.MapKeyColumn;
+import jakarta.persistence.Column;
+import jakarta.persistence.JoinColumn;
+
+
+import lombok.Getter;
+import lombok.Setter;
+import lombok.ToString;
+import java.io.Serializable;
 import java.util.Map;
 import java.util.Set;
 
-public class Effect {
+@Entity
+@Getter
+@Setter
+@ToString
+public class Effect implements Serializable {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
     private String type;
     private String target;
     private int value;
     private String action;
-    private Map<String, Object> condition;
     private int amount;
-    private String operation; 
+    private String operation;
+
+    @ElementCollection
     private Set<Keyword> keywords;
     private String triggerCondition;
 
-    // Getters and Setters
-    public String getType() {
-        return type;
-    }
+    @ElementCollection
+    @CollectionTable(name = "effect_conditions", joinColumns = @JoinColumn(name = "effect_id"))
+    @MapKeyColumn(name = "condition_key")
+    @Column(name = "condition_value")
+    private Map<String, String> condition;
 
-    public void setType(String type) {
-        this.type = type;
-    }
+    @ManyToOne
+    @JoinColumn(name = "card_name")
+    private Card card;
 
-    public String getTarget() {
-        return target;
-    }
-
-    public void setTarget(String target) {
-        this.target = target;
-    }
-
-    public int getValue() {
-        return value;
-    }
-
-    public void setValue(int value) {
-        this.value = value;
-    }
-
-    public String getAction() {
-        return action;
-    }
-
-    public void setAction(String action) {
-        this.action = action;
-    }
-
-    public Map<String, Object> getCondition() {
-        return condition;
-    }
-
-    public void setCondition(Map<String, Object> condition) {
-        this.condition = condition;
-    }
-
-    public int getAmount() {
-        return amount;
-    }
-
-    public void setAmount(int amount) {
-        this.amount = amount;
-    }
-
-    public String getOperation() {
-        return operation;
-    }
-
-    public void setOperation(String operation) {
-        this.operation = operation;
-    }
-
-    public Set<Keyword> getKeywords() {
-        return keywords;
-    }
-
-    public void setKeywords(Set<Keyword> keyword) {
-        this.keywords = keyword;
-    }
-
-    public String getTriggerCondition() {
-        return triggerCondition;
-    }
-
-    public void setTriggerCondition(String triggerCondition) {
-        this.triggerCondition = triggerCondition;
-    }
 }
