+++ mindbug-backend/src/main/java/com/mindbug/services/GameSessionFactory.java
@@ -19,7 +19,7 @@ public class GameSessionFactory {
         GameSession gameSession = this.applicationContext.getBean(GameSession.class, game, 
         this.applicationContext.getBean(WSMessageManager.class),
         this.applicationContext.getBean(GameSessionValidation.class), this.applicationContext,
-        this.applicationContext.getBean(PlayerService.class), this.applicationContext.getBean(GameService.class));
+        this.applicationContext.getBean(PlayerService.class));
         return gameSession;
     }
 }
